Refactor functions to exist inside a class
